requirements.txt
amqp==2.5.2
ansicon==1.89.0
APScheduler==3.6.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.1
Django==3.0.2
django-apscheduler==0.3.0
django-cors-headers==3.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
greenlet==0.4.15
importlib-metadata==1.5.0
jdcal==1.4.1
jinxed==1.0.0
kombu==4.6.8
monotonic==1.5
mysqlclient==1.4.6
openpyxl==3.0.3
Pillow==7.0.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
tzlocal==2.0.0
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
